feat: Track objects spawned by SpawnSyncedObject components during generation
These spawned objects don't get attached to the parent that spawned them, so we have to keep them in a dictionary to reference when preparing the tile contents for culling.

A `TileContentsBuilder` class is added to allow concatenating the contents of different objects.

`TileContents` will now not be guaranteed to only contain the children of the `Tile` they are associated with. This allows doors to be included on their two adjacent `TileContents`, which prevents the doorways next to the camera from disappearing when the tile they are associated with is culled.

Tile contents are prepared by `DungeonCullingInfo` instead now, since it will later have to run through all lights in the scene to calculate their influence on each tile.