feat: Keep external lights' influence on tiles if culling them will be visible
By calculating which lights influence which tiles, we can avoid any static lights popping into or out of view when culling changes.

Lights with shadows disabled as well as lights that don't fully dim their light in shadow will be added to the TileContents based on the range to the light.

For lights that have shadows enabled, any renderers that may cast shadows into the light cast into a tile will be added to the TileContents as well. Any light that is fully dimmed by being in shadow will be added based on line of sight through portals to the light, so that they will be culled if they are fully occluded from casting light on the tile in question.

Note that this doesn't take into account light shape yet, so that is an optimization that may improve performance in the future. For example, spot lights will be treated as if they are point lights and be checked in a sphere instead of a cone.