optimization: Pre-calculate tile bounds into TileContents
The `DunGen.Tile.Bounds` calculates the bounds of the tile based on the transform of the tile on every call. We should be able to assume that tiles never move. By pre-calculating the bounds at generation time, finding the tile a camera is in, as well as finding all tiles visible to an orthographic camera, will waste less time on this transform step.

In the process, `DungeonCullingInfo` has been refactored not to include an array of `DunGen.Tile` instances. When calculating light influences, keeping the tile a light is in will no longer be checked using `GetTile(this Vector3)`, since they are retrieved from their containing tile already.

Yields an approximate 3.8% increase in framerate upon entering the fire exit on Titan seed 98506861.