fix: Increase the allowed radius outside a tile to 10
This fixes an issue where the camera would not be considered inside Scoopys Variety Mod's sewer interior entrance tile if standing to the right of the door. Note that this is a workaround to an issue with the actual entrance tile's bounds that can also allow tiles to clip inside the entrance room in the same place where culling fails. Once this is fixed, the radius can be reduced again.