fix: Make teleportation tracking more reliable
Use the `PlayerControllerB.TeleportPlayer()` to track teleports that don't go through an entrance.

Hook into `EntranceTeleport.TeleportPlayer()` which runs for the local player as well as the existing `EntranceTeleport.TeleportPlayerClientRpc()` function which runs for other clients in the game.

By making these changes, the last `PlayerControllerB.isInsideFactory` value seen will actually be reliable in vanilla. Mods may break this assumption that it will be set correctly after the last hook that runs, but that's something we will have to discover later.

Far clip plane adjustments will now be applied from the start of the game regardless of the culling method used.