feat: Make use of new culling info for the depth culling implementation
We have this information for the portal occlusion culling, but it's also usable for depth culling as well, so let's switch over to that. This simplifies the depth culling code quite a bit.

This will also fix most issues where the depth culling would hide the tile that the player is standing in. It can still happen if a tile's bounds are wildly inaccurate.