feat: Use `LateUpdate()` to apply portal occlusion culling
Using `LateUpdate()` has the following benefits:
- Any calls to the `Renderer.isVisible` property should be accurate according to whether it was culled in the previous frame.
- Any errors that occur while applying the culling should not take down the rendering pipeline, and screen freezing should no longer occur.

It has the downside of individual cameras having to render more, i.e. when the player is looking at the map screen, the entire area it is viewing will be visible to the gameplay camera and any body cams. However this appears to have a negligible effect on framerate in practice.