feat: Only add light occluders for lights that are in line of sight
Previously, a lot of light occluders would be added to a tile when they didn't actually occlude the light.

Any tiles that contained lights within range of another tile would be added to that tile's light occluders list, regardless of whether the light was an influence.

All renderers in a tile were added as light occluders regardless of whether they were in line of sight of the light. Now, only ones that are within the portals that a light passes through will be included, further reducing the number of light occluders that need to be enabled for a given tile.