fix: Cull lights using the `cullingMask` field instead of `enabled`
Apparently something has been overriding the Light components to always remain enabled. It doesn't appear to be anything in a script, since `Behaviour.set_enabled()` is never called on a light to reset it.

However, using the `cullingMask` instead should help prevent conflicts with other mods, since the field is not even present in the editor. The original values are kept to be properly restored, but that may not be totally necessary, as I didn't observe any lights actually using a mask other than `0xFFFF'FFFF`.

This fixes cases in which some static lights would shine through walls. The performance difference appears to be negligible.