feat: Only check light contribution for active objects
There are a lot of light prefabs that get spawned but are never active in the mansion interior, so let's check that they're active before calculating their line of sight and occluders.

This will mean that any lights that become active after the generation completes may get culled incorrectly, but generally that isn't a thing that should happen. If mods start doing this, then I guess we'll have to hook in and update the culling information when that happens, but that would not be ideal.