feat: Cull dynamic lights and items
Culling dynamic lights allows us to keep more tiles culled as well, due to the fact that light can be stopped from shining through walls.

In addition, this will cull items on the surface while in the interior and vice versa, to avoid them contributing to lower framerates when looking in certain directions in late game.