feat: Calculate the bounds of grabbable objects only once
It's not necessary to calculate the bounds every time we check for visibility or intersection, since those bounds will not change during a culling call. Instead, cache them when they are used and then reuse those bounds for the rest of the culling.

Cuts the time spent collecting visible items from ~160 microseconds down to ~7 microseconds. Total time in culling went from a very high >400 microseconds down to about 150-160 microseconds.