feat: Only warn about exceeding the maximum portal stack depth once
The fixed depth of 16 is almost never exceeded in vanilla interiors. In instances where it was, the missing tiles were never visible, so it should be fine to just print a warning once and then stop.

If we find that this happens consistently and visibly on a custom dungeon, then the capacity can definitely be increased, the memory footprint is minimal regardless.