optimization: Directly refer to TileContents in portals
This allows us to skip using dictionaries to convert between DunGen classes and our tile and portal classes when doing visibility testing.

In the process, also ditch the `TileContentsBuilder` class, which only makes it more difficult to do visibility testing when we're still in the process of building the `TileContents` instances.