fix: Use scan nodes to determine if items are in the dungeon
These should be the most accurate indicator of where an item is, and we can fall back to the exact position of the `GrabbableObject` itself if the scan node doesn't exist, instead of using the renderer bounds.

This fixes an issue where shotguns grabbed from the corpses of nutcrackers would be invisible when dropped outside the dungeon, which was caused by a `ParticleSystemRenderer` that remained within the dungeon's bounds after leaving it.