fix: Exclude getting camera position from benchmark
It seems that getting the camera position when the player is no longer parented to the ship results in us measuring the time to cache the world position of the camera. This measures a whole 60 microseconds on my machine, but there doesn't appear to be any way to avoid that slowdown. Let's ignore this so that we can get more useful measurements of the rest.