fix: Reduce light occluders with a normal plane for portal frustums
Since the frustum didn't have a normal plane to prevent it from considering things on the opposite side of the frustum to be within it, dynamic lights were often keeping tiles visible outside the camera completely unnecessarily.

Caching the plane means that this should only really affect the number of planes that must be tested, but that cost is negligible in practice.