feat: Add an option to disable lights' shadow distance fading
Lights in the Mansion interior like to have their shadow fade distance set to the same value as their range, meaning that as soon as you go further than twice the radius of the light, it will start shining through walls. This causes the chandelier lights in the entry tile to be visible through the walls of neighboring tiles, and breaks shadowed light influence checks in DungeonCullingInfo.

In the interest of keeping the optimizations that these light influence checks provide while also making the lighting more correct, I'm adding an option that is defaulted to `true` that prevents such lights from shining through walls, so that they and the tiles they reside in may be culled when they are blocked by a wall.