fix: Switch back to Awake() and OnDestroy() for CullingMethod.Instance
Turns out that we hide all the lights if the Instance isn't set to null before OnDynamicLightsCollected() is called. However, we were never setting it to null after c7704aa.

Instead use DestroyImmediate() to prevent the issue that commit was intended to fix, and set the Instance in Awake() and clear it in OnDestroy().