fix: Don't replace the initial far plane values when options change
The backup of the initial value of the player cameras' far planes could be stored multiple times if a mod like LethalConfig changed the values in-game. This would cause setting the CullDistanceEnabled option to false to reset it to the previous value that was set with the override instead of its vanilla value.