fix: Increase the shadow fade distance to avoid flickers when fading
If the shadows begin fading at the same time that the light itself begins fading, then the light becomes partially visible through the wall before it fully fades. This happens most visibly on the extremely bright lights shining from the fireplaces in the mansion interior.

To counter this, scale the shadow fade distance based on the observed point at which it begins fading, which appears to be at 90% of the fade distance.