refactor: Keep config migration constants where they are used
Storing these in fields decouples the old config values from the condition that indicates which version they were in. By using constants inline in the migration code, one can simply look at the version number in the condition, and then see on the next line what values it looks for in the config entries to modify.