feat: Use a Burst plugin to test the intersection of planes and an AABB
The built-in Unity runtime method for this, GeometryUtility.TestPlanesAABB, gets stuck in an infinite loop if it iterates to a plane at the index 32.

By implementing this in Burst, we can get similar speeds to the Unity runtime without this downside.

The method is implemented by transposing the operations that would be performed on each corner of the AABB, so that the dot products can be calculated in parallel using SIMD.