optimization: Disable lights' shadow fading based a heuristic
This should resolve a performance regression when running portal occlusion culling on the Grand Armory interior. The shadow fading distance of its lights generally caused the shadows to be invisible when shadow fading was allowed, but when it was disallowed, the shadows would render far more often, causing a severe degradation of performance.