fix: Test dynamic object visibility by the bounds of renderers in tiles
Previously, we were using a radius around the actual tile bounds to determine the playable area of a tile that we should consider for culling dynamic objects like items and lights.

Instead, when we create the TileContents, create a bounding box containing all the renderers in the tile, with an upper limit of 20 units outside the tile to prevent it expanding to the exterior surface. This bounding box is considered the tentatively playable area of the tile, so that entry tiles on the vanilla mineshaft and Scoopy's sewer with their larger upper rooms will be culled properly when no tiles intersect.