fix: Calculate item positions from targetFloorPosition in local space
When an item is spawned using SpawnSyncedObject in the interior, it will be parented to MapProps, and the position will be set relative to that. Therefore, we need to apply the parent transform to the targetFloorPosition to get its actual landing position to determine whether it is in the interior.