fix: Actually use the item contents comparer for item visibility sets
The identity comparer doesn't suffice here, because GrabbableObjectContents is just a wrapper around the GrabbableObject component. The fact that the map of items to their contents is refreshed also causes there to be multiple instances, so the comparer is especially needed.

Fixes an issue where items were initially invisible due to the contents wrappers being refreshed and the new ones not being considered to be visible in the existing item visibility sets.