fix: Make sure refreshed GrabbableObjectContents are placed in the sets
Because GrabbableObjectContents has a comparison based on the underlying GrabbableObject, when adding to either AllGrabbableObjectContentsInInterior or AllGrabbableObjectsOutside, the instance of GrabbableObjectContents wasn't necessarily up-to-date. To fix this, we just do a little extra work by removing the previous instance in all cases, then re-add it to whatever set it should be in after the refresh. This has the extra benefit of making RefreshGrabbableObject() look a little cleaner.