fix: Account for lights that lie outside their parent tile
To prevent these from behaving differently with culling on, we must treat them as always visible to tiles they affect. We must also make any shadow casters between the light and those tiles as visible, in order to keep the shadows consistent when they pass through walls. Without portals, there's no way to determine when we've fully occluded the light, so we just assume that everything in its path needs to be visible.

This adds a new method that creates a minimal convex hull around an axis-aligned bounding box and a single point, in order to detect as small a number of occluders as possible.