fix: Use the automatic tile bounds calculation for overridden ones too
Apparently the method of calculating the world bounds of tiles with overridden bounding boxes was incorrect in tiles with scale. It appears that the existing calculation that was used for automatic tile bounds also works in those cases, so we can just cut that code and use one method for both.